MAINT: deduplicate out/casting in ufuncs
