+++ torch_np/__init__.py
@@ -4,3 +4,4 @@ from . import testing
 
 from ._unary_ufuncs import *
 from ._binary_ufuncs import *
+from ._ndarray import can_cast, result_type
